Fix autoupdate to not switch to completely different tags
If multiple tags to the release of a hook `pre-commit autoupdate` would
previously pick the first tag, possibly non-deterministically
depending on a locally configure Git `tag.sort` configuration. If
multiple tags point to the release hook release, e.g., if multiple
projects are tracked in the same repo, this could cause switching to a
different projects tag. At the very least this introduces churn and
makes it hard to see what version was upgraded to, and worst case could
cause `pre-commit autoupdate` behave differently on different machines.

With this patch we always prefer to pick the tag which shares the longest
prefix with a currently configured tag.

Closes pre-commit/pre-commit#3349.